# swagger.yml
basePath: /
consumes:
- application/json
host: localhost:9098
info:
  description: Dictionary Service to store and retrieve JSON like hierarchical dictionaries
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Dictionary Service to store and retrieve JSON like hierarchical dictionaries
  version: 0.0.18
paths:
...
  /api/calendar/{calendarType}/{from}/{to}:
    get:
      description: Loads all calendar items for given type and date range
      operationId: Calendar
      parameters:
      - in: path
        name: CalendarType
        required: true
        type: string
      - format: date-time
        in: path
        name: DayFrom
        required: true
        type: string
      - format: date-time
        in: path
        name: DayTo
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/loadCalendar'
        "400":
          $ref: '#/responses/RestError'
        default:
          $ref: '#/responses/RestError'
...
  /api/metadata:
    post:
      description: Create new dictionary type
      operationId: createMetadata
      parameters:
      - description: Dictionary type key
        in: body
        name: Type
        schema:
          type: string
      - description: Json Schema describes dictionary type
        in: body
        name: Content
        schema:
          type: string
      responses:
        "200":
          description: ""
        "400":
          $ref: '#/responses/RestError'
produces:
- application/json
responses:
  RestError:
    description: RestError general service error message
    headers:
      error:
        description: Error message
        type: string
      error_description:
        description: An optional field describes error in more details
        type: string
  loadCalendar:
    description: ""
    headers:
      day:
        description: Day in YYYY-MM-dd format
        type: string
      kind:
        description: Describe day kind in calendar
        type: string
      labels:
        description: Optional extra labels
      name:
        description: |-
          Day name in calendar
          required: true
        type: string
      tenant:
        type: string
    schema:
      additionalProperties:
        type: string
      type: object
...
schemes:
- http
swagger: "2.0"